LibGfx/JPEGXL: Use u32 for group numbers
This should be enough for quite some time and allows us to use the
values in initializer lists without `static_cast`s.

+++ Userland/Libraries/LibGfx/ImageFormats/JPEGXLLoader.cpp
@@ -2517,8 +2517,8 @@ struct Frame {
     u64 width {};
     u64 height {};
 
-    u64 num_groups {};
-    u64 num_lf_groups {};
+    u32 num_groups {};
+    u32 num_lf_groups {};
 
     Optional<Image> image {};
 };
